Move astar and pacman wrapper to separate namespace
