+++ demo/test/mock_environment_model.hpp
@@ -37,7 +37,7 @@ public:
     template <std::size_t Size>
     void setMaze(const Position& size, const char (&str)[Size]) {
         mazeState_ = makeCustomMazeState({size.x, size.y}, str);
-        astar_ = AStar(mazeState_);
+        astar_ = utils::AStar(mazeState_);
     }
     void setEmptyMaze() {
         const char str[] = {"##########"
@@ -54,7 +54,7 @@ public:
     }
 
 private:
-    Game dummyGame_; // We just need this to initialize the base environment model
+    entt::Game dummyGame_; // We just need this to initialize the base environment model
 };
 
 } // namespace demo
